Task1/consumer/requirements.txt
pika
py-postgresql
